- New example on how to use yaocptool.parallel.Worker. It is used for solving an NLP with ADMM, which each step is executed in a different process.
- New method for Discretization Schemes: unpack_decision_variables,
which get the solution of the NLP and separate it into states,
algebraics, controls, and other decision variables.

- .plot can take 'all' instead of a list of indices to be plotted. 'all'
makes that all variables of the the given types is plotted
eg: .plot([{'x':'all}])

- OCP now can have optimization parameter (ocp.p_opt), which are time
invariant optimization variables.

- MultipleShooting now also include the algebraic variables in the NLP.
This allows to put bound and other constraints on y.

- SimulationResults has a new method .extend, for extending a simulation
 result. Useful for keeping a chain of simulation in a single object.

- New Classes in yaocptool.mpc: MPC, Plant,PlantSimulation.

- parallel.worker: can be stopped by the object, takes a list of input
queues and put the result in a list of output queues.

- new module yaocptool.stochastic, with some function related to PCE
(Polynomial Chaos Expansion)